# photo-prestige/openapi/source.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Photo Prestige API Gateway",
    "version": "1.2.0",
    "description": "Complete gateway documentation for all available API endpoints.\\n\\nRole model: participant, target-owner.\\n\\nEach operation includes role access in the description and x-allowed-roles."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local gateway"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AuthPayload": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "role": {
            "type": "string",
            "enum": ["participant", "target-owner"],
            "description": "Required on register, ignored on login."
          }
        }
      },
      "TargetCreate": {
        "type": "object",
        "required": ["title", "imageUrl", "lat", "lng", "deadlineAt"],
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "imageUrl": { "type": "string", "format": "uri" },
          "locationDescription": { "type": "string" },
          "city": { "type": "string" },
          "lat": { "type": "number" },
          "lng": { "type": "number" },
          "radiusMeters": { "type": "number", "example": 250 },
          "deadlineAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "SubmissionCreate": {
        "type": "object",
        "required": ["imageUrl"],
        "properties": {
          "imageUrl": { "type": "string", "format": "uri" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register user",
        "description": "Allowed roles: public.",
        "x-allowed-roles": ["public"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthPayload" }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered" },
          "400": { "description": "Validation error" },
          "409": { "description": "User already exists" }
        }
      }
    },
...
    "/api/v1/targets": {
      "get": {
        "tags": ["Targets"],
        "summary": "List targets",
        "description": "Allowed roles: public.",
        "x-allowed-roles": ["public"],
        "parameters": [
          { "name": "city", "in": "query", "schema": { "type": "string" } },
          { "name": "ownerId", "in": "query", "schema": { "type": "string" } },
          { "name": "active", "in": "query", "schema": { "type": "boolean" } },
          { "name": "lat", "in": "query", "schema": { "type": "number" } },
          { "name": "lng", "in": "query", "schema": { "type": "number" } },
          { "name": "radiusMeters", "in": "query", "schema": { "type": "number" } },
          { "name": "radiusKm", "in": "query", "schema": { "type": "number" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "skip", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Targets list" },
          "400": { "description": "Invalid query" }
        }
      },
      "post": {
        "tags": ["Targets"],
        "summary": "Create target",
        "description": "Allowed roles: target-owner.",
        "x-allowed-roles": ["target-owner"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TargetCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "Target created" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
...
    "/api/v1/targets/{targetId}/submissions": {
      "post": {
        "tags": ["Score"],
        "summary": "Create submission",
        "description": "Allowed roles: participant, target-owner.",
        "x-allowed-roles": ["participant", "target-owner"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SubmissionCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "Submission created" },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Target not found" },
          "409": { "description": "Target closed" }
        }
      }
    },
...
    "/api/v1/targets/{targetId}/score": {
      "get": {
        "tags": ["Score"],
        "summary": "Get my best score",
        "description": "Allowed roles: participant, target-owner.",
        "x-allowed-roles": ["participant", "target-owner"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Best score" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "No score found" }
        }
      }
    },
    "/api/v1/targets/{targetId}/scores": {
      "get": {
        "tags": ["Score"],
        "summary": "Get all scores for target",
        "description": "Allowed roles: target-owner (must own target).",
        "x-allowed-roles": ["target-owner"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "All scores" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Target not found" }
        }
      }
    },
    "/api/v1/targets/{targetId}/winner": {
      "get": {
        "tags": ["Score"],
        "summary": "Get winner for target",
        "description": "Allowed roles: participant, target-owner.",
        "x-allowed-roles": ["participant", "target-owner"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Winner" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Winner not calculated yet" }
        }
      }
    },
    "/api/v1/targets/{targetId}/registrations": {
      "post": {
        "tags": ["Register"],
        "summary": "Create registration",
        "description": "Allowed roles: participant, target-owner.",
        "x-allowed-roles": ["participant", "target-owner"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "201": { "description": "Registration created" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Target not found" },
          "409": { "description": "Registration closed or duplicate" }
        }
      }
    },
    "/api/v1/targets/{targetId}/participants": {
      "get": {
        "tags": ["Register"],
        "summary": "List participants for target",
        "description": "Allowed roles: target-owner (must own target).",
        "x-allowed-roles": ["target-owner"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Participants list" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Target not found" }
        }
      }
    },
    "/api/v1/me/registrations": {
      "get": {
        "tags": ["Register"],
        "summary": "List my registrations",
        "description": "Allowed roles: participant, target-owner.",
        "x-allowed-roles": ["participant", "target-owner"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "My registrations" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/v1/targets/{targetId}/registrations/me": {
      "delete": {
        "tags": ["Register"],
        "summary": "Cancel my registration for target",
        "description": "Allowed roles: participant, target-owner.",
        "x-allowed-roles": ["participant", "target-owner"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Registration cancelled" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Registration not found" }
        }
      }
    }
  }
}